+++ .github/workflows/codspeed.yml
@@ -1,0 +1,46 @@
+---
+name: codspeed
+
+on:
+    push:
+        branches:
+            - master
+    pull_request:
+        paths:
+            - w3lib/**
+            - tests/benchmarks/**
+            - .github/workflows/codspeed.yml
+            - pyproject.toml
+    workflow_dispatch:
+
+concurrency:
+    group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+    cancel-in-progress: true
+
+permissions:
+    contents: read
+
+jobs:
+    build:
+        runs-on: ubuntu-latest
+        strategy:
+            fail-fast: false
+            matrix:
+                python-version: ['3.13', pypy3.11]
+
+        steps:
+            - uses: actions/checkout@v6
+
+            - name: Set up Python ${{ matrix.python-version }}
+              uses: actions/setup-python@v6
+              with:
+                  python-version: ${{ matrix.python-version }}
+
+            - name: Run benchmarks on ${{ matrix.python-version }}
+              uses: CodSpeedHQ/action@v4
+              with:
+                  pytho~n-version: ${{ matrix.python-version }}
+                  run: |
+                      pip install --upgrade pip
+                      pip install --upgrade tox
+                      tox -e benchmark-${{ matrix.python-version }}
