+++ tox.ini
@@ -10,10 +10,11 @@ envlist = py310, py311, py312, py313, py314, pypy3.11, docs, pylint, typing, pre
 deps =
     pytest !=3.1.1, !=3.1.2
     pytest-cov >= 7.0.0
+    pytest-xdist
 commands =
     python -m pytest \
         --doctest-modules \
-        --cov=w3lib --cov-report=term --cov-report=xml \
+        --cov=w3lib --cov-report=term --cov-report=xml -n auto \
         {posargs:w3lib tests}
 
 [testenv:typing]
@@ -51,22 +52,18 @@ commands =
     python -m build --sdist
     twine check dist/*
 
-[testenv:benchmark-{py39,py310,py311,py312,py313,py314,pypy310,pypy311}]
+[testenv:benchmark-{py313,pypy311}]
 basepython =
-    py39: python3.9
-    py310: python3.10
-    py311: python3.11
-    py312: python3.12
     py313: python3.13
-    py314: python3.14
-    pypy310: pypy3.10
     pypy311: pypy3.11
 deps =
-    pytest
+    {[testenv]deps}
     pytest-codspeed
 commands =
-    python -m pytest \
+    python -m pytest tests/benchmarks \
+        --no-cov \
         --codspeed \
         --codspeed-warmup-time=1 \
         --codspeed-max-rounds=10000 \
-        --codspeed-max-time=10
+        --codspeed-max-time=10 \
+        -n auto
